LayoutTests/http/tests/w3c/webperf/submission/user-timing/test_user_timing_method_exist-expected.txt

Description

This test validates that all of the user timing methods exist in window.performance (but does not validate that their functionality).


PASS window.performance is defined 
PASS performance.webkitMark is supported! 
PASS performance.webkitClearMarks is supported! 
PASS performance.webkitMeasure is supported! 
PASS performance.webkitClearMeasures is supported! 
PASS performance.webkitGetEntries is supported! 
PASS performance.webkitGetEntriesByType is supported! 
PASS performance.webkitGetEntriesByName is supported! 

